refactor(action): fold CaptureSnapshotIos into CaptureSnapshot platform switch
Merge the Android and iOS snapshot-capture commands into a single
CaptureSnapshot class that dispatches on device.platform via a switch,
with private executeAndroid/executeIos methods and Ios-prefixed iOS
helpers (issue #5461). Replaces the platform branch in
deviceSnapshotManager.createCaptureProvider with a single construction;
the Android-only constructor guard is dropped in favor of the switch.

Behavior-preserving. iOS coverage moves into CaptureSnapshot.test.ts and
the SnapshotProvider conformance case targets the folded class.
